chore(refactor): separate module for `RetcherOptions`
