+++ src/lib.rs
@@ -3,51 +3,12 @@ use napi_derive::napi;
 
 mod response;
 mod request;
+mod retcher_builder;
 
+use request::{HttpMethod, RequestInit};
+use retcher_builder::RetcherOptions;
 use self::response::RetchResponse;
 
-#[napi(string_enum)]
-pub enum Browser {
-  Chrome,
-  Firefox,
-}
-
-impl Into<retch::Browser> for Browser {
-  fn into(self) -> retch::Browser {
-    match self {
-      Browser::Chrome => retch::Browser::Chrome,
-      Browser::Firefox => retch::Browser::Firefox,
-    }
-  }
-}
-
-#[derive(Default)]
-#[napi(object)]
-struct RetcherOptions {
-  pub browser: Option<Browser>,
-  pub ignore_tls_errors: Option<bool>,
-  pub vanilla_fallback: Option<bool>,
-  pub proxy_url: Option<String>,
-}
-
-impl Into<RetcherConfig> for RetcherOptions {
-  fn into(self) -> RetcherConfig {
-    let mut config = RetcherConfig::default();
-    if let Some(browser) = self.browser {
-      config = config.with_browser(browser.into());
-    }
-    if let Some(ignore_tls_errors) = self.ignore_tls_errors {
-      config = config.with_ignore_tls_errors(ignore_tls_errors);
-    }
-    if let Some(vanilla_fallback) = self.vanilla_fallback {
-      config = config.with_fallback_to_vanilla(vanilla_fallback);
-    }
-    if let Some(proxy_url) = self.proxy_url {
-      config = config.with_proxy(proxy_url);
-    }
-    config
-  }
-}
 
 #[napi(js_name="Retcher")]
 pub struct RetcherWrapper {
@@ -59,9 +20,10 @@ impl RetcherWrapper {
     #[napi(constructor)]
     pub fn new(options: Option<RetcherOptions>) -> Self {
       let config: RetcherConfig = options.unwrap_or_default().into();
-        Self {
-          inner: config.build(),
-        }
+
+      Self {
+        inner: config.build(),
+      }
     }
 
     #[napi]
